requirements.txt
bitstring==3.1.7
future==0.18.2
lxml==4.6.1
numpy==1.19.2
pygame==2.0.0.dev6
pymavlink==2.4.12
pyserial==3.4
